rev 232: fix mapping
fa16b0675144ef2143b1d5fd3a92807a57b40e7f 7ce6cc2716e138fc58db28921069d089d5b364c7

+++ runescape-api/src/main/java/net/runelite/rs/api/RSIntProjection.java
@@ -40,29 +40,29 @@ public interface RSIntProjection extends IntProjection
 
 	@Import("pitchSin")
 	@Override
-	int getPitchSin();
+	float getPitchSin();
 
 	@Import("pitchSin")
-	void setPitchSin(int pitchSin);
+	void setPitchSin(float pitchSin);
 
 	@Import("pitchCos")
 	@Override
-	int getPitchCos();
+	float getPitchCos();
 
 	@Import("pitchCos")
-	void setPitchCos(int pitchCos);
+	void setPitchCos(float pitchCos);
 
 	@Import("yawSin")
 	@Override
-	int getYawSin();
+	float getYawSin();
 
 	@Import("yawSin")
-	void setYawSin(int yawSin);
+	void setYawSin(float yawSin);
 
 	@Import("yawCos")
 	@Override
-	int getYawCos();
+	float getYawCos();
 
 	@Import("yawCos")
-	void setYawCos(int yawCos);
+	void setYawCos(float yawCos);
 }
